requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-multipart
oracledb